apps/api-services/requirements.txt

# Git packages
-r requirements/git.txt

# Async HTTP
aiohttp>=3.13.3

# Password Hashing
argon2-cffi>=25.1.0

# AWS
boto3>=1.42.47

# Web Scraping
exa-py>=2.4.0

# Core Web Framework
fastapi==0.129.0

# HTTP Clients
httpx==0.28.1
inflection==0.5.1

# Templates & Forms
jinja2>=3.1.6

# Kafka
kafka-python-ng>=2.2.3
litellm>=1.81.10

# LiveKit (Voice/Video)
livekit-api>=1.1.0
mistralai==1.5.1

# Microservices
nameko>=2.14.1

# AI & LLM
openai>=2.20.0

# Data Processing
pandas>=2.2.3
psycopg2-binary>=2.9.11
pycryptodome>=3.23.0

# Data Validation
pydantic==2.12.5

# Authentication & Security
PyJWT[crypto]>=2.11.0

# Testing
pytest>=9.0.2
pytest-asyncio>=1.3.0
pytest-mock>=3.15.1
python-dateutil>=2.9.0
python-dotenv>=1.2.1
python-multipart>=0.0.22
pytz>=2025.2

# Utilities
PyYAML>=6.0.3
redis[hiredis]>=7.1.1
rsa>=4.9
uvicorn[standard]==0.40.0
